- Add `BusyStateHelper`, `ErrorStateHelper` to `BaseViewModel` - Add all the `BuilderHelpers` into one mixin
